+++ src/Symfony/Component/Form/AbstractType.php
@@ -20,8 +20,9 @@ use Symfony\Component\OptionsResolver\OptionsResolverInterface;
 abstract class AbstractType implements FormTypeInterface
 {
     /**
-     * The extensions for this type
-     * @var array An array of FormTypeExtensionInterface instances
+     * @var array
+     *
+     * @deprecated Deprecated since version 2.1, to be removed in 2.3.
      */
     private $extensions = array();
 
@@ -44,14 +45,6 @@ abstract class AbstractType implements FormTypeInterface
      */
     public function finishView(FormViewInterface $view, FormInterface $form, array $options)
     {
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    public function createBuilder($name, FormFactoryInterface $factory, array $options)
-    {
-        return null;
     }
 
     /**
@@ -98,21 +91,26 @@ abstract class AbstractType implements FormTypeInterface
     }
 
     /**
-     * {@inheritdoc}
+     * Sets the extensions for this type.
+     *
+     * @param array $extensions An array of FormTypeExtensionInterface
+     *
+     * @throws Exception\UnexpectedTypeException if any extension does not implement FormTypeExtensionInterface
+     *
+     * @deprecated Deprecated since version 2.1, to be removed in 2.3.
      */
     public function setExtensions(array $extensions)
     {
-        foreach ($extensions as $extension) {
-            if (!$extension instanceof FormTypeExtensionInterface) {
-                throw new UnexpectedTypeException($extension, 'Symfony\Component\Form\FormTypeExtensionInterface');
-            }
-        }
-
         $this->extensions = $extensions;
     }
 
     /**
-     * {@inheritdoc}
+     * Returns the extensions associated with this type.
+     *
+     * @return array An array of FormTypeExtensionInterface
+     *
+     * @deprecated Deprecated since version 2.1, to be removed in 2.3. Use
+     *             {@link ResolvedFormTypeInterface::getTypeExtensions()} instead.
      */
     public function getExtensions()
     {