+++ src/Symfony/Component/Form/CHANGELOG.md
@@ -86,6 +86,7 @@ CHANGELOG
    * `hasAttribute`
    * `getClientData`
  * added FormBuilder methods
+   * `getTypes`
    * `addViewTransformer`
    * `getViewTransformers`
    * `resetViewTransformers`
@@ -157,3 +158,15 @@ CHANGELOG
  * deprecated the options "data_timezone" and "user_timezone" in DateType, DateTimeType and TimeType
    and renamed them to "model_timezone" and "view_timezone"
  * fixed: TransformationFailedExceptions thrown in the model transformer are now caught by the form
+ * added FormRegistry and ResolvedFormTypeInterface
+ * deprecated FormFactory methods
+   * `addType`
+   * `hasType`
+   * `getType`
+ * [BC BREAK] FormFactory now expects a FormRegistryInterface as constructor argument
+ * [BC BREAK] The method `createBuilder` in FormTypeInterface is not supported anymore for performance reasons
+ * [BC BREAK] Removed `setTypes` from FormBuilder
+ * deprecated AbstractType methods
+   * `getExtensions`
+   * `setExtensions`
+ * ChoiceType now caches its created choice lists to improve performance